Moved code to follow gb folder structure.
